tpx3: allow to cancel while reading header
`stream_recv_header` allows to pass in a closure that can, for example,
check for control messages. This allows to cancel if the code is
currently waiting for a header to be sent. Before, it would just hang in
an EAGAIN loop.

With this change, the upstream code also needed to be special cased to
not reconnect if the acquisition was cancelled.